Extract static HTML transcript renderer
