+++ docs/DECISIONS.md
@@ -174,6 +174,7 @@
 - Static HTML review rendering lives in `WorkspaceHTMLReviewRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates Git review pane, file, hunk, line, inline comment, and action markup while staying responsible for transcript placement.
 - Static HTML sidebar rendering lives in `WorkspaceHTMLSidebarRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates project rows, thread groups, bulk-selection controls, primary actions, and the tools/settings footer while staying responsible for whole-workspace shell composition.
 - Static HTML top-bar rendering lives in `WorkspaceHTMLTopBarRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates model/mode, status, runtime issue pill, Computer Use status, and overflow command markup while staying responsible for whole-workspace shell composition.
+- Static HTML transcript rendering lives in `WorkspaceHTMLTranscriptRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates transcript empty state, context banner, runtime issue panel, review placement, message timeline rows, tool-card row handoff, and composer markup while staying responsible only for composing the whole static harness document.
 - Browser state and presentation contracts live in `QuillCodeBrowserSurface.swift`. `BrowserState`, `BrowserSnapshotState`, `BrowserCommentState`, `BrowserSurface`, `BrowserSnapshotSurface`, and `BrowserCommentSurface` evolve together; browser address resolution lives in `WorkspaceBrowserLocationResolver.swift`; and browser history/page/comment state transitions live in `WorkspaceBrowserEngine`. The workspace model owns async fetch orchestration, `lastError`, and top-bar refreshes, while reusable browser state contracts, URL-resolution rules, and pure state transitions stay focused and guarded by parity tests.
 - Review comment payload state and event planning live in `WorkspaceReviewCommentPlanner.swift`, not in `WorkspaceModel.swift`. The planner owns path/text trimming, visible review-file validation, line-range normalization, line-kind checks, summary copy, and payload encoding. The workspace model owns only selected-thread validation, event append, persistence, and top-bar refresh.
 - Agent tool execution override composition lives in `WorkspaceToolExecutionOverrideCombiner.swift`, not in `WorkspaceModel.swift`. The combiner owns the precedence order Plan, Remote Project, Browser, Computer Use, Memory, MCP, while the workspace model only creates each optional executor and assigns the combined override to the runner.
